# api/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "auth_internal_users.GetUserResponse": {
        "properties": {
          "created_at": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "email_verified": {
            "type": "boolean"
          },
          "id": {
            "type": "string"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "updated_at": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "users.ClientParams": {
        "properties": {
          "allowed_scopes": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "name": {
            "type": "string"
          },
          "redirect_uri": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "users.ClientResponse": {
        "properties": {
          "allowed_scopes": {
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "created_at": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "redirect_uri": {
            "type": "string"
          },
          "revoked_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          },
          "user_id": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "users.DeleteUserParams": {
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "users.ListClientsResponse": {
        "properties": {
          "clients": {
            "items": {
              "$ref": "#/components/schemas/users.ClientResponse"
            },
            "type": "array"
          },
          "next_cursor": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "users.UpdatePasswordParams": {
        "properties": {
          "current_password": {
            "type": "string"
          },
          "new_password": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "users.UpdateUserParams": {
        "properties": {
          "email": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "users.UpdateUserResponse": {
        "properties": {
          "email_verification_required": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "description": "Type \"Bearer\" followed by a space and JWT token.",
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "api for auth.hisbaan.com",
    "title": "auth",
    "version": "0.0.1"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/users/me": {
      "delete": {
        "description": "Deletes the authenticated user's account",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/users.DeleteUserParams"
              }
            }
          },
          "description": "Account deletion confirmation",
          "required": true,
          "x-originalParamName": "request"
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Delete current user",
        "tags": [
          "users"
        ]
      },
      "get": {
        "description": "Returns the authenticated user's profile",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/auth_internal_users.GetUserResponse"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Get current user",
        "tags": [
          "users"
        ]
      },
      "put": {
        "description": "Updates the authenticated user's profile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/users.UpdateUserParams"
              }
            }
          },
          "description": "User update details",
          "required": true,
          "x-originalParamName": "request"
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/users.UpdateUserResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Update current user",
        "tags": [
          "users"
        ]
      }
    },
    "/users/me/clients": {
      "get": {
        "description": "Returns a paginated list of OAuth clients for the authenticated user",
        "parameters": [
          {
            "description": "Pagination cursor",
            "in": "query",
            "name": "cursor",
            "schema": {
              "type": "string"
            }
          },
          {
            "description": "Maximum number of results (max 100)",
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/users.ListClientsResponse"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "List current user's clients",
        "tags": [
          "users"
        ]
      },
      "post": {
        "description": "Creates an OAuth client for the authenticated user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/users.ClientParams"
              }
            }
          },
          "description": "Client details",
          "required": true,
          "x-originalParamName": "request"
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Create client",
        "tags": [
          "users"
        ]
      }
    },
    "/users/me/clients/{clientId}": {
      "delete": {
        "description": "Deletes an OAuth client owned by the authenticated user",
        "parameters": [
          {
            "description": "Client ID",
            "in": "path",
            "name": "clientId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Delete client",
        "tags": [
          "users"
        ]
      },
      "put": {
        "description": "Updates an OAuth client owned by the authenticated user",
        "parameters": [
          {
            "description": "Client ID",
            "in": "path",
            "name": "clientId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/users.ClientParams"
              }
            }
          },
          "description": "Updated client details",
          "required": true,
          "x-originalParamName": "request"
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Update client",
        "tags": [
          "users"
        ]
      }
    },
    "/users/me/clients/{clientId}/revoke": {
      "post": {
        "description": "Revokes an OAuth client owned by the authenticated user",
        "parameters": [
          {
            "description": "Client ID",
            "in": "path",
            "name": "clientId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Revoke client",
        "tags": [
          "users"
        ]
      }
    },
    "/users/me/password": {
      "post": {
        "description": "Changes the authenticated user's password",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/users.UpdatePasswordParams"
              }
            }
          },
          "description": "Password change details",
          "required": true,
          "x-originalParamName": "request"
        },
        "responses": {
          "204": {
            "description": "No Content"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Change password",
        "tags": [
          "users"
        ]
      }
    }
  }
}